Initial experiment.
Already an interesting learning experience.

Used `cargo-generate` with the template https://github.com/rp-rs/rp2040-project-template
which sets up a lot of the fiddly config bits correctly.

One wrinkle is the default project code blinks the LED on the regular Pi
Pico board.  It doesn't work on Pi Pico W where the GPIO pin is somehow
not connected to the on-board LED.  I didn't figure out how to set the
on board LED initially.  But I modified the code to use GPIO pin 15 and
rigged up a breadboard circuit to flash a separate LED and that works.

Using the debug probe, it's as easy as `cargo run`, and the firmware is
compiled and flashed to the board and immediately executed.