Get the button reading API working.
Seems to be pretty reliable, but the mapping of the physical buttons on
the board to the row number is not intuitive.  It seems the right-most
four are mapped to even numbered rows, and the left-most four to
odd-numbered rows, all of them in column 1.

Either that or I have a bug in the decode logic.  That needs to be
looked into.